TIP-871: Move compiler pass to the right name space
